requirements.txt
bibtexparser==1.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
crossref-commons==0.0.7
doi2bib==0.4.0
future==0.18.3
gender-guesser==0.4.0
idna==3.6
install==1.3.5
langdetect==1.0.9
mysql-connector==2.2.9
neo4j==5.16.0
pyparsing==3.1.1
pytz==2023.3.post1
ratelimit==2.2.1
requests==2.31.0
six==1.16.0
urllib3==2.1.0
